% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_tas}
\alias{county_tas}
\title{sf POLYGON object of local government areas on the main Tasmanian island}
\format{
An sf object with 27 features (rows) and 4 fields (variables):
\describe{
\item{country}{\url{http://rs.tdwg.org/dwc/terms/country}}
\item{countryCode}{\url{http://rs.tdwg.org/dwc/terms/countryCode}}
\item{stateProvince}{\url{http://rs.tdwg.org/dwc/terms/stateProvince}}
\item{county}{\url{http://rs.tdwg.org/dwc/terms/county}}
\item{geometry}{simple feature geometries}
}
}
\usage{
county_tas
}
\description{
Polygons and labels are modified from shape data provided by the Australian
Bureau of Statistics showing local government areas (the Australian
Darwin Core 'county' equivalent). Note that the polygons have been simplified
to reduce the file size. The data is for example use only.
}
\keyword{datasets}
