% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crs_data}
\alias{crs_data}
\title{Data for semi-major axis and ellipsoid flattening for EPSG codes}
\format{
A data frame with 869 rows and 4 variables:
\describe{
\item{epsg_code}{EPSG code for the coordinate reference system}
\item{name}{Name of the coordinate reference system}
\item{semi_major_axis}{radius at the equator for the given datum}
\item{inverse_flattening}{inverse of the flattening component of the datum}
}
}
\usage{
crs_data
}
\description{
Calculating coordinate precision uncertainty requires the values for the
semi-major axis and the ellipsoid flattening for the specified coordinate
reference system. These data were generated using \code{\link[sf]{st_crs}} to
extract the the semi-major axis and ellipsoid flattening values from the EPSG
codes.
}
\keyword{datasets}
