% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_coordinatePrecision.R
\name{dwc_coordinatePrecision}
\alias{dwc_coordinatePrecision}
\title{Calculate the coordinatePrecision in degrees for geographic coordinates}
\usage{
dwc_coordinatePrecision(
  verbatimLongitude,
  verbatimLatitude,
  precision = NULL,
  unit = NULL
)
}
\arguments{
\item{verbatimLongitude}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/verbatimLongitude}.}

\item{verbatimLatitude}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/verbatimLatitude}.}

\item{precision}{The precision of the coordinates. Only provide a value if
verbatimLongitude and verbatimLatitude are in "degrees minutes seconds" or
"degrees decimal minutes", or if you wish to override the function and
specify the precision for all coordinates. Default is NULL.}

\item{unit}{The unit of precision (one of either "degrees", "minutes" or
"seconds"). Only provide a value if precision is given.}
}
\value{
A numeric. A value that can be used in the Darwin Core coordinatePrecision
field. See: \url{http://rs.tdwg.org/dwc/terms/coordinatePrecision}. The
maximum precision returned is 0.00001 degrees, or approximately 1.11 m at the
equator.
}
\description{
If verbatimLongitude and verbatimLatitude are in decimal degrees, this
function will use the coordinate values to calculate coordinate precision.
Alternatively, the coordinate precision can be assigned by specifying both
the precision and unit arguments. The precision and unit arguments must be
supplied if the verbatim coordinates are in "degrees minutes seconds" or
"degrees decimal minutes".
}
\examples{
dwc_coordinatePrecision(144.9786, -41.17472)
}
