% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_coordinates.R
\name{dwc_coordinates}
\alias{dwc_coordinates}
\title{Prepare geographic coordinates and convert to decimal degrees if needed}
\usage{
dwc_coordinates(longitude, latitude, verbatimCoordinateSystem, verbatimSRS)
}
\arguments{
\item{longitude}{The longitude of the given coordinates This will be assigned
to \href{http://rs.tdwg.org/dwc/terms/verbatimLongitude}{verbatimLongitude}
in the returned tibble. Must be a character value if the coordinates are in
"degrees decimal minutes" or "degrees minutes seconds".}

\item{latitude}{The latitude of the given coordinates This will be assigned
to \href{http://rs.tdwg.org/dwc/terms/verbatimLatitude}{verbatimLatitude}
in the returned tibble. Must be a character value if the coordinates are in
"degrees decimal minutes" or "degrees minutes seconds".}

\item{verbatimCoordinateSystem}{The format of the longitude and latitude
coordinates. Supported terms are "decimal degrees", "degrees decimal minutes"
or "degrees minutes seconds". See:
\url{http://rs.tdwg.org/dwc/terms/verbatimCoordinateSystem}}

\item{verbatimSRS}{The spatial reference system associated with the
coordinates. Given as an EPSG code as in the example: "EPSG:4326". See:
\url{http://rs.tdwg.org/dwc/terms/verbatimSRS}.}
}
\value{
A \code{\link[tibble]{tibble}} with Darwin Core terms
\href{http://rs.tdwg.org/dwc/terms/decimalLatitude}{decimalLatitude},
\href{http://rs.tdwg.org/dwc/terms/decimalLongitude}{decimalLongitude},
\href{http://rs.tdwg.org/dwc/terms/geodeticDatum}{geodeticDatum},
\href{http://rs.tdwg.org/dwc/terms/verbatimLatitude}{verbatimLatitude},
\href{http://rs.tdwg.org/dwc/terms/verbatimLongitude}{verbatimLongitude},
\href{http://rs.tdwg.org/dwc/terms/verbatimCoordinateSystem}{verbatimCoordinateSystem}
and \href{http://rs.tdwg.org/dwc/terms/verbatimSRS}{verbatimSRS}.
}
\description{
This function prepares the raw (verbatim) coordinate information and if
necessary, converts coordinates from 'degrees minutes seconds' or 'degrees
decimal minutes' to 'decimal degrees' using the R package
\href{https://docs.ropensci.org/parzer/}{parzer}.Where longitude and latitude are
provided in decimal degrees, this function will round these to a maximum of
seven decimal places.
}
\examples{
data("thylacine_data")
thylacine_data |>
dplyr::mutate(dwc_coordinates(longitude = longitude_dms, latitude = latitude_dms,
verbatimCoordinateSystem = "degrees minutes seconds",
verbatimSRS = "EPSG:4326"))
}
