% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_coordinateUncertaintyInMeters.R
\name{dwc_coordinateUncertaintyInMeters}
\alias{dwc_coordinateUncertaintyInMeters}
\title{Calculate the coordinateUncertaintyInMeters}
\usage{
dwc_coordinateUncertaintyInMeters(
  decimalLatitude,
  coordinatePrecision,
  geodeticDatum,
  gps_uncertainty = 30
)
}
\arguments{
\item{decimalLatitude}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
Latitude in decimal degrees. See: \url{http://rs.tdwg.org/dwc/terms/decimalLatitude}.}

\item{coordinatePrecision}{Output of \code{\link[dwcPrepare]{dwc_coordinatePrecision}}.
The precision of the coordinates in degrees. Value must be between 0.00001
and 1. See: \url{http://rs.tdwg.org/dwc/terms/coordinatePrecision}.}

\item{geodeticDatum}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}. The
spatial reference system associated with the coordinates. Given as an EPSG
code as in the example: "EPSG:4326". See: \url{http://rs.tdwg.org/dwc/terms/geodeticDatum}.}

\item{gps_uncertainty}{The uncertainty in meters recorded by the GPS device.
Default is 30.}
}
\value{
A numeric. A value that can be used in the Darwin Core
\href{http://rs.tdwg.org/dwc/terms/coordinateUncertaintyInMeters}{coordinateUncertaintyInMeters}
field.
}
\description{
This function calculates the uncertainty in meters associated with a given
set of geographic coordinates. The uncertainty is the sum of two parts:
\enumerate{
\item the uncertainty associated with the coordinate precision given a datum and
latitude. This is calculated following the equations found at:
\url{http://georeferencing.org/georefcalculator/docs/GeorefGuide.html#imprecision_in_coordinates}.
If the coordinate reference system is unknown, the default value of 5,359
meters is assigned for this component. See Details.
\item the uncertainty associated with the GPS recording device. By default this
is given as 30 meters.
}
}
\details{
Please see, Wieczorek C, J Wieczorek (2021) Georeferencing Calculator.
Available:  \url{http://georeferencing.org/georefcalculator/gc.html} for more
options and further details.

The uncertainty associated with geographic coordinates from an unknown datum
is taken from
\url{https://docs.gbif-uat.org/georeferencing-best-practices/1.0/en/#uncertainty-from-unknown-datum}.
Note that this is a worst-case scenario whilst this function lacks the
functionality to correctly determine the uncertainty of an unknown datum. If
you have only a small number of locations and an unknown datum, we recommend
using the Wieczorek C, J Wieczorek (2021) Georeferencing Calculator. The
output of the calculator can be copied into R using
\code{\link[dwcPrepare]{dwc_format_gco}}.
}
\examples{
dwc_coordinateUncertaintyInMeters(-41.17472, 0.01, "EPSG:4326", 30)
}
