% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_country_to_county.R
\name{dwc_country_to_county}
\alias{dwc_country_to_county}
\title{Assign country, countryCode, stateProvince and county to point locations}
\usage{
dwc_country_to_county(
  df,
  decimalLongitude,
  decimalLatitude,
  county_sf,
  country_column_name = "country",
  countryCode_column_name = "countryCode",
  stateProvince_column_name = "stateProvince",
  county_column_name = "county"
)
}
\arguments{
\item{df}{A data frame containing the decimal longitude and latitude
coordinates}

\item{decimalLongitude}{The longitude of the focal point in decimal degrees.
Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/decimalLongitude}.}

\item{decimalLatitude}{The latitude of the focal point in decimal degrees.
Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/decimalLatitude}.}

\item{county_sf}{An \code{sf POLYGON} object that includes the county
boundaries along with the higher geography features.}

\item{country_column_name}{The column name in \code{county_sf} that gives the
name of the country in which the county is found. The default value is
"country".}

\item{countryCode_column_name}{The column name in \code{county_sf} that gives
the countryCode of the country in which the county is found. The default
value is "countryCode".}

\item{stateProvince_column_name}{The column name in \code{county_sf} that
gives the State or Province in which the county is found.The default value is
"stateProvince".}

\item{county_column_name}{The column name in \code{county_sf} that gives the
county name. The default value is "county".}
}
\value{
The original \code{df} object with the additional columns
\href{http://rs.tdwg.org/dwc/terms/country}{country},
\href{http://rs.tdwg.org/dwc/terms/countryCode}{countryCode},
\href{http://rs.tdwg.org/dwc/terms/stateProvince}{stateProvince} and
\href{http://rs.tdwg.org/dwc/terms/county}{county}.
}
\description{
Takes point locations as an \code{sf} POINT object and county boundaries as
an \code{sf} POLYGON object and uses \code{\link[sf]{st_join}} to assign
country, countryCode, stateProvince and county Darwin Core terms to each
point location.
}
\details{
Note that the function expects county to be nested within stateProvince which
is in turn nested within country. The \code{sf} POLYGON object with the
county boundaries should contain features describing these higher level
geographies.
}
\examples{
# Make up some point data
point_locations <- data.frame(
  name = c("A", "B", "C"),
  lon = c(117.093225, 127.502052, 115.674972),
  lat = c(-33.110168, -25.128663, -33.982473))

# Make up some county polygon data nested within higher geographies
county_sf <-
  sf::st_buffer(sf::st_as_sf(point_locations,
                             coords = c("lon", "lat"),
                             remove = FALSE,
                             crs = sf::st_crs("EPSG:4326")), dist = 1000) |>
  dplyr::mutate(
    country_raw = c("Country 1", "Country 2", "Country 3"),
    countryCode_raw = c("C1", "C2", "C3"),
    stateProvince_raw = c("State 1", "State 2", "State 3"),
    county_raw = c("County 1", "County 2", "County 3")
  )

# Run the function
dwc_country_to_county(point_locations,
                    decimalLongitude = "lon",
                    decimalLatitude = "lat",
                    county_sf = county_sf,
                    country_column_name = "country_raw",
                    countryCode_column_name = "countryCode_raw",
                    stateProvince_column_name = "stateProvince_raw",
                    county_column_name = "county_raw")
}
