% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_Event.R
\name{dwc_Event}
\alias{dwc_Event}
\title{Prepare Darwin Core Event fields}
\usage{
dwc_Event(
  start,
  end = NA,
  time_NA_value = "00:00:00",
  tzone = "UTC",
  sampleSizeUnit = "day",
  fieldNumber = NA,
  habitat = NA,
  samplingProtocol = NA,
  samplingEffort = NA,
  fieldNotes = NA,
  eventRemarks = NA
)
}
\arguments{
\item{start}{The start date or date-time of the sampling event. Accepts
character strings with dates in one of the following formats: "dmy HMS"
(eg, "24/12/2012 15:45:00"), "ymd HMS" (eg, "2012-12-24 15:45:00"), "dmy"
(eg, "24/12/2012" or "ymd" (eg, "2012-12-24").}

\item{end}{The end date or date-time of the sampling event. Default is NA.}

\item{time_NA_value}{Where a mixture of dates/date-times occur, the dates
without times are often assigned a time of 00:00:00. This is the default
value here, and is used so that the eventDate field is shown correctly. We
suggest being very cautious if changing the default.}

\item{tzone}{The timezone of the date-time passed to
\code{\link[lubridate]{parse_date_time}}. The default is "UTC".}

\item{sampleSizeUnit}{The unit of time for calculating the length of the
event. Passed to \code{\link[lubridate]{time_length}} as the \code{unit}
argument. The default is "day".}

\item{fieldNumber}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/fieldNumber}{fieldNumber} term. Default is
NA.}

\item{habitat}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/habitat}{habitat} term. Default is NA.}

\item{samplingProtocol}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/samplingProtocol}{samplingProtocol} term.
Default is NA.}

\item{samplingEffort}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/samplingEffort}{samplingEffort} term.
Default is NA.}

\item{fieldNotes}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/fieldNotes}{fieldNotes} term. Default is
NA.}

\item{eventRemarks}{Optional argument for what should be included in the
\href{http://rs.tdwg.org/dwc/terms/eventRemarks}{eventRemarks} term. Default
is NA.}
}
\value{
A \code{\link[tibble]{tibble}} with Darwin Core terms
\href{http://rs.tdwg.org/dwc/terms/fieldNumber}{fieldNumber} (if supplied),
\href{http://rs.tdwg.org/dwc/terms/eventDate}{eventDate},
\href{http://rs.tdwg.org/dwc/terms/startDayOfYear}{startDayOfYear},
\href{http://rs.tdwg.org/dwc/terms/endDayOfYear}{endDayOfYear},
\href{http://rs.tdwg.org/dwc/terms/year}{year},
\href{http://rs.tdwg.org/dwc/terms/month}{month},
\href{http://rs.tdwg.org/dwc/terms/day}{day},
\href{http://rs.tdwg.org/dwc/terms/verbatimEventDate}{verbatimEventDate},
\href{http://rs.tdwg.org/dwc/terms/habitat}{habitat} (if supplied),
\href{http://rs.tdwg.org/dwc/terms/samplingProtocol}{samplingProtocol}
(if supplied),
\href{http://rs.tdwg.org/dwc/terms/sampleSizeValue}{sampleSizeValue} (here
given as the length of time of the event, given if both the \code{start} and
\code{end} are given),
\href{http://rs.tdwg.org/dwc/terms/sampleSizeUnit}{sampleSizeUnit} (given if
both the \code{start} and \code{end} are given),
\href{http://rs.tdwg.org/dwc/terms/samplingEffort}{samplingEffort} (if
supplied),
\href{http://rs.tdwg.org/dwc/terms/fieldNotes}{fieldNotes} (if supplied) and
\href{http://rs.tdwg.org/dwc/terms/eventRemarks}{eventRemarks} (if supplied).
}
\description{
This function takes a start date/date-time as well as an optional end
date/date-time as its main arguments and produces the Darwin Core Event
terms. If both a start and end are given the length of time the event took
place is assigned to the sampleSizeValue field. If only one of the start
or end values have a time component, both values are treated as dates.
\code{dwc_Event} is essentially a wrapper for a series of
\href{https://lubridate.tidyverse.org/}{lubridate} functions, and experienced R
users may prefer to use the lubridate functions directly.
}
\examples{
dwc_Event(start = "2012-12-24 15:00:00", end = "2012-12-27 16:00:00")

# The \code{\link[tibble]{tibble}} will be added to an existing dataframe
# if combined with dplyr's mutate

library("dplyr")

# Load data
data("thylacine_data")

thylacine_data |>
  mutate(dwc_Event(start = date_trap_setup,
                   end = date_trap_collected))

}
