% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_format_gco.R
\name{dwc_format_gco}
\alias{dwc_format_gco}
\title{Format the output of \url{http://georeferencing.org/georefcalculator/gc.html}}
\usage{
dwc_format_gco(location_identifier, georeference_calculator_output)
}
\arguments{
\item{location_identifier}{The name of the location. Used so that the
returned \code{\link[tibble]{tibble}} can be joined with other data.}

\item{georeference_calculator_output}{The output of the georeferencing
calculator at \url{http://georeferencing.org/georefcalculator/gc.html}.}
}
\value{
A \code{\link[tibble]{tibble}}. Column names are Darwin Core terms.
}
\description{
The online \href{http://georeferencing.org/georefcalculator/gc.html}{georeferencingcalculator}
allows for the results to be output by selecting 'Copy' and then
pasted into a spreadsheet. This function takes the output as an argument and
formats it as a \code{\link[tibble]{tibble}}.
}
\examples{
# Note that the georeference_calculator_output has been truncated for this
# example
dwc_format_gco(
  location_identifier = "A",
  georeference_calculator_output =
    "10.27	-123.6	WGS84	1557	0.0000001	JB	2022-08-31	Georeferencing..."
)
}
