% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_locality.R
\name{dwc_locality}
\alias{dwc_locality}
\title{Locality descriptor for given longitude/latitude coordinates}
\usage{
dwc_locality(
  decimalLongitude,
  decimalLatitude,
  localityPointLocations_sf = dwcPrepare:::locality_data_aus,
  locationName = "localityName",
  ...
)
}
\arguments{
\item{decimalLongitude}{The longitude of the focal point}

\item{decimalLatitude}{The latitude of the focal point}

\item{localityPointLocations_sf}{An \code{sf} POINT object. Points must be in
longitude / latitude. If no object is given the function uses internal data
that contains Australian locality information.}

\item{locationName}{The column name in
\code{localityPointLocations_sf} that gives the location name. If using the
default data this should be left at the default value "localityName".}

\item{...}{Additional arguments passed to \code{\link[sf]{st_distance}}}
}
\value{
A string that could be used in the Darwin Core locality field,
\url{https://dwc.tdwg.org/terms/#location}.
}
\description{
This function finds the closest location from a set of candidate locations to
the given longitude/latitude, as well as the name of the closest location,
and the direction of the closest location from the focal longitude/latitude.
}
\examples{
localityPointLocations_sf <-
sf::st_as_sf(data.frame(
 name = c("A", "B", "C"),
 lon = c(117.093225, 127.502052, 115.674972),
 lat = c(-33.110168, -25.128663, -33.982473)),
 coords = c("lon", "lat"),
 crs = sf::st_crs("EPSG:4326"))


dwc_locality(
 decimalLongitude = 115.949465, decimalLatitude = -32.029655,
 localityPointLocations_sf = localityPointLocations_sf, locationName = "name"
 )

}
