% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_point_cluster.R
\name{dwc_point_cluster}
\alias{dwc_point_cluster}
\title{Cluster a set of geographic coordinates (longitude, latitude)}
\usage{
dwc_point_cluster(
  df,
  longitude,
  latitude,
  distm_method = geosphere::distHaversine,
  hclust_method = "average",
  k = NULL,
  h = NULL
)
}
\arguments{
\item{df}{A data frame object.}

\item{longitude}{The name of the column in \code{df} containing the longitude
values.}

\item{latitude}{The name of the column in \code{df} containing the latitude
values.}

\item{distm_method}{The method passed to \code{\link[geosphere]{distm}} for
calculating the distance between points.}

\item{hclust_method}{The method passed to \code{\link[stats]{hclust}} for
performing the hierarchical clustering.}

\item{k}{The number of desired groups passed to \code{\link[stats]{cutree}}.}

\item{h}{The distance cutoff passed to \code{\link[stats]{cutree}}.}
}
\value{
A data frame with a new column \code{cluster} that shows the cluster each
point has been assigned to.
}
\description{
This function uses \code{\link[geosphere]{distm}} to create a distance matrix
and \code{\link[stats]{hclust}} with \code{\link[stats]{cutree}} to assign
each point to a cluster.
}
\examples{
df <-
data.frame(
longitude = c(116.018, 116.010, 116.011, 115.969, 115.976, 115.973),
latitude = c(-31.709, -31.707, -31.713, -32.339, -32.339, -32.344)
)

dwc_point_cluster(df, "longitude", "latitude",
                   hclust_method = "average",  h = 5000)
}
