% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_polygon_cluster.R
\name{dwc_polygon_cluster}
\alias{dwc_polygon_cluster}
\title{Find polygons (sites) that intersect}
\usage{
dwc_polygon_cluster(x)
}
\arguments{
\item{x}{An \code{sf} object that contains the site locations as polygons.}
}
\value{
\code{x} with the additional column \code{.cluster}.
}
\description{
This function identifies polygons that intersect in some way. Its main use is
to identify sites within a data set that cover the same or similar area, and
could be considered a single site. A cluster identifuer (\code{.cluster}) is
assigned to each polygon, with intersecting polygons identified by a common
\code{.cluster} id.
}
\examples{
data("thylacine_data")

# Generate some interseting polygons
thylacine_polygons <-
  thylacine_data |>
  dplyr::group_by(trap) |>
  dplyr::slice(1) |>
  sf::st_as_sf(coords = c("longitude_dd", "latitude_dd"), crs = "EPSG:4326") |>
  sf::st_buffer(dist = 5000) |>
  dplyr::select(site, trap, geometry) |>
  dplyr::ungroup()

# Run function
dwc_polygon_cluster(thylacine_polygons)
}
