% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwcPrepare-package.R
\docType{package}
\name{dwcPrepare-package}
\alias{dwcPrepare}
\alias{dwcPrepare-package}
\title{dwcPrepare: Helps Prepare Data for use in a Darwin Core Archive}
\description{
The package helps take a data frame that contains data such as geographic coordinates, date/date-time and species abundances, and transforms this data ready for use in a Darwin Core archive.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jdberson/dwcPrepare}
  \item Report bugs at \url{https://github.com/jdberson/dwcPrepare/issues}
}

}
\author{
\strong{Maintainer}: Jacob Berson \email{jacob.berson@uwa.edu.au} (\href{https://orcid.org/0000-0002-7059-2131}{ORCID})

Authors:
\itemize{
  \item Jake Manger \email{jake.manger@uwa.edu.au} (\href{https://orcid.org/0000-0002-0686-4788}{ORCID})
}

}
\keyword{internal}
