% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locality_data_aus}
\alias{locality_data_aus}
\title{sf POINT object of Australian localities}
\format{
An sf object with 16197 features (rows) and 1 fields (variables):
\describe{
\item{locality_name}{name of the locality}
\item{geometry}{simple feature geometries}
}
}
\usage{
locality_data_aus
}
\description{
Data is from "Elvis - Place Names - Foundation Spatial Data - Locality Data",
sourced from Commonwealth of Australia (The Intergovernmental Committee on
Surveying and Mapping, ICSM) 2018, under a Creative Commons Attribution 4.0
International license. See \url{https://placenames.fsdf.org.au/}. Note that
data received has been subset so that only the NAME field and geometry
column are retained in the sf object. NAME has been renamed to locality_name.
}
\keyword{datasets}
