% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thylacine_data}
\alias{thylacine_data}
\title{A made-up survey data set for use in examples and vignettes}
\format{
A data frame with 8 rows and 13 variables:
\describe{
\item{site}{name given to each site}
\item{trap}{there were two traps at each site, 1 and 2}
\item{date_trap_setup}{the date time the trap was setup}
\item{date_trap_collected}{the date time the trap was collected}
\item{longitude_dd}{longitude of the trap in decimal degrees}
\item{latitude_dd}{latitude of the trap in decimal degrees}
\item{longitude_dms}{longitude of the trap in degrees minutes seconds}
\item{latitude_dms}{latitude of the trap in degrees minutes seconds}
\item{longitude_ddm}{longitude of the trap in degrees decimal minutes}
\item{latitude_ddm}{latitude of the trap in degrees decimal minutes}
\item{gps_uncertainty}{the uncertainty of the coordinates given by the GPS}
\item{species}{the species being identified / counted}
\item{count}{the number of individuals found in a trap}
}
}
\usage{
thylacine_data
}
\description{
Tasmanian Tigers (Thylacinus cynocephalus) are rumoured to still roam the
wilds of Tasmania. This is a 'mock' dataset of a survey of Tasmanian Tigers.
Sadly no tigers were caught!
}
\keyword{datasets}
