% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_coordinatePrecision.R
\name{dwc_coordinatePrecision}
\alias{dwc_coordinatePrecision}
\title{Calculate the coordinatePrecision in degrees for geographic coordinates}
\usage{
dwc_coordinatePrecision(
  verbatimLongitude,
  verbatimLatitude,
  verbatimCoordinateSystem
)
}
\arguments{
\item{verbatimLongitude}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/verbatimLongitude}.}

\item{verbatimLatitude}{Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/verbatimLatitude}.}

\item{verbatimCoordinateSystem}{The format of the longitude and latitude
coordinates. Supported terms are "decimal degrees", "degrees decimal minutes"
or "degrees minutes seconds". See:
\url{http://rs.tdwg.org/dwc/terms/verbatimCoordinateSystem}}
}
\value{
A numeric. A value that can be used in the Darwin Core coordinatePrecision
field. See: \url{http://rs.tdwg.org/dwc/terms/coordinatePrecision}. The
maximum precision returned is 0.00001 degrees, or approximately 1.11 m at the
equator.
}
\description{
This function calculates the precision of geographic coordinates.
}
\examples{
# For a single point
dwc_coordinatePrecision(144.9786, -41.17472,
verbatimCoordinateSystem = "decimal degrees")

# For a set of points
data("thylacine_data")
thylacine_data |>
dplyr::mutate(coordinatePrecision =
dwc_coordinatePrecision(verbatimLatitude = latitude_dd,
                        verbatimLongitude = longitude_dd,
                        verbatimCoordinateSystem = "decimal degrees"))

}
