% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwc_locality.R
\name{dwc_locality}
\alias{dwc_locality}
\title{Locality descriptor for given longitude/latitude coordinates}
\usage{
dwc_locality(
  decimalLongitude,
  decimalLatitude,
  localities_sf,
  localities_names,
  ...
)
}
\arguments{
\item{decimalLongitude}{The longitude of the focal point in decimal degrees.
Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/decimalLongitude}.}

\item{decimalLatitude}{The latitude of the focal point in decimal degrees.
Output of \code{\link[dwcPrepare]{dwc_coordinates}}.
See: \url{http://rs.tdwg.org/dwc/terms/decimalLatitude}.}

\item{localities_sf}{An \code{sf POINT} object that provides
locality names as well as their longitude and latitude for all localities
within the area of interest. The package includes the locality information
for Australia. See \code{\link[dwcPrepare]{locality_data_aus}}}

\item{localities_names}{The column name in \code{localities_sf} that gives
the locality name.}

\item{...}{Additional arguments passed to \code{\link[sf]{st_distance}}}
}
\value{
A string that could be used in the Darwin Core locality field,
\url{http://rs.tdwg.org/dwc/terms/locality}.
}
\description{
This function finds the closest locality to the given point location, as well
as the distance and direction of the closest locality from the focal
location. The returned string could be used in the Darwin Core
\href{http://rs.tdwg.org/dwc/terms/locality}{locality} field.
}
\examples{
localities_sf <-
sf::st_as_sf(data.frame(
 name = c("A", "B", "C"),
 lon = c(117.093225, 127.502052, 115.674972),
 lat = c(-33.110168, -25.128663, -33.982473)),
 coords = c("lon", "lat"),
 crs = sf::st_crs("EPSG:4326"))


dwc_locality(
 decimalLongitude = 115.949465, decimalLatitude = -32.029655,
 localities_sf = localities_sf, localities_names = "name"
 )

}
